refactor(tests): update sign-in method and improve brand consistency tests
- Replaced `signInDoctorAndSetCookies` with `signInDoctorOrSkipOnInfraError` in multiple test files to enhance error handling during sign-in.
- Updated landing page tests to check for new heading content, ensuring alignment with recent branding changes.
- Adjusted navigation and public site availability tests to reflect updated expected text for improved consistency across tests.
